BAU: Fix Jest errors in IDE
VSCode couldn't find the Jest types, so was giving us errors in all the
test files.

This commit adds Jest to the `tsconfig.json` and removes the test files
from the exclude. This solves the IDE problems but means we're including
the test files in the complied Javascript.

That's not ideal either, but we're a way off deploying these to
production so we'll fix it later.

https://stackoverflow.com/questions/54139158/cannot-find-name-describe-do-you-need-to-install-type-definitions-for-a-test
Signed-off-by: divyaparameswaran <divya.parameswaran@digital.cabinet-office.gov.uk>