GUA-426: Write parsed events to DynamoDB
This commit sets up the 'happy path' for writing the incoming events to
DynamoDB.

The DynamoDB client is set up outside the handler function - this allows
it to be reused if the function is called from a warm start, gaining us
a bit of performance.

The `userServices` object is logged to the console at the moment; this
will need to be removed before we deploy to production to prevent
leaking potentially sensitive data.