GUA-426: Refactor handler function
Separate out parsing and validation into their own functions and put
them inside the try / catch clause.

This allows each function to throw a relevant error and they'll all get
caught and handled in the same way by the top-level function. This will
be useful in the future when we want to put all Records that cause an
error onto a dead letter queue.