GUA-426: Move validation into main file
Based on some feedback from Paul J:

> And try to keep everything to one file.
The point being that having more than one file is completely unnecessary for something that only needs 200 lines of code to work, so why split it up?
The purpose of creating objects, classes, methods etc originally was so that these could be shared across multiple files. In a Lambda function, this is not needed.
Also, if you know that most of your functions are single files, you know where most of your code is.
For typescript, having types in a separate file would make sense I
suggest since that isn't going to be transpiled to js

I've left the models and error files out for now as those will be moved
out to a shared library in the future. We might move the validation
logic as well, but that's not so certain.