BAU: Don't error when dev dependencies are imported in tests
Add a rule config which allows us to import from development
dependencies in test files.

Without this, `import { mockClient } from "aws-sdk-client-mock";` was
raising a linting error for extraneous dependencies[1] because the
library is a development dependency.

That's ok here though, because the import is in a test file and won't
get run when the lambda is deployed.

[1]: https://github.com/import-js/eslint-plugin-import/blob/v2.26.0/docs/rules/no-extraneous-dependencies.md