GUA-426: Remove custom `Error` types
We weren't doing anything with the `ValidationError`s that couldn't be
done with the base `Error` class.

`getErrorMessage` was also superfluous - `console.error()` attempts to
cast its parameters to a `string` anyway, so passing the `Error` object
is safe.

This means we can get rid of the error file and tests completely.