GUA-248: Set all DeletionPolicy and UpdateReplacePolicy to Delete
cfn-lint tells us that the implicit default policy here is Delete, which
is undertandably risky given we're talking about persistence that could
be on prod.

It's recommendation is to make this explicit.

The initial bet was to vary things based on which environment we were
in, Deleting on our dev instance for a clean workspace, and retaining on
prod.

Our if statements are not playing nicely with sam (`sam validate`).
So we'll need to come up with a new plan.

For now, let's go with delete and then work out a plan for prod when
we're closer to that