GUA-436: Add action to test lambdas
This will run when a pull request is opened and run linting and the unit
tests for each lambda.

It's possible that this actions ends up taking a long time to run,
particularly if we end up with a lot of lambdas in this repo, or if some
of those lambdas have a lot of dependencies.

If that happens we can look at splitting this action up or caching
dependencies[1].

[1]:https://docs.github.com/en/actions/using-workflows/caching-dependencies-to-speed-up-workflows