GUA-427: Add linting config
Note we've turned off the rule import/no-extraneous-dependencies for any
test files. When bringing adding `import "aws-sdk-client-mock-jest`[1] to
the top of a test file we were triggering an annoying error saying the
library should be in dependencies not devDependencies. That's clearly
wrong.

So this configures this to target the rule only to our non-test
typescript

[1]: https://github.com/m-radzikowski/aws-sdk-client-mock#jest-matchers

Co-authored-by: Divya Parameswaran <divya.parameswaran@digital.cabinet-office.gov.uk>
Co-authored-by: Huw Diprose <huw.diprose@digital.cabinet-office.gov.uk>