GUA-448: Resolve promises in handler function
Using `.forEach` with an `async` callback function means we generate a
`Promise` for each item in the array, but we never wait for those
promises to resolve (or reject) before the lambda reaches the end of the
handler and stops running.

This means the records never get processed and nothing is put on the
queue for the next lambda. Switch to using `.map` on the array which
returns an array of promises, then wait for them to resolve before
stopping the lambda function.

[This Stack Overflow answer](https://stackoverflow.com/questions/37576685/using-async-await-with-a-foreach-loop/37576787#37576787)
explains the problem better than I can.