GUA-448: Make data structure consistent
The query lambda outputs a message with key `TxmaEvent` but this one was
expecting the key to be `TxmaEventBody`. Update this lambda to match the
query lambda output.

Also make the same switch to wait for all the promises to resolve as
added in fa668d5361b78748fcc87f583d8f0333b30d0c2e