GUA-448: Correct data type stored in DynamoDB
This lambda fetched a user's data from DynamoDB and cast it to type
`Service[]`, but it's actually `UserServices`. Because we're casting
from an `any` type in the response from DynamoDB, this meant all the
type checking and tests passed here but the format lambda failed as the
data type in the message it received was wrong.

Change this lambda to expect the correct data type from the DynamoDB
query, and pass through just the services list.