GUA-448: Create permissions policy before role
We need to create the permissions policy before the roles for each
lambda are created, otherwise it fails to deploy with an error that the
lambda functions don't have permission to read events off the queues
that trigger them.

Formatting the file like this has no impact on what ends up deployed,
but ensures that the policy is created before the roles.