GUA-431: Pass a string to messageBodyString instead of object
`object` is a bit anonymous, we looked at type definition for a while
and then deicded that we don't really need to parse this out of JSON
string. The dead letter queue can just pass on things naievely and then
it's down to our error handling to work out what to make of payloads.