GUA-438: Exclude test files from deployed Lambdas
We removed the `exclude` directive for the test files in #2 to get
VSCode to recognise that tests were run with Jest and not show errors
for unknown functions.

Further testing has shown we only need the extra `types` entry added
there, and VSCode is happy when the test files are added back to the
exclusion list.

This change doesn't seem to have reduced the size of the deployed
Lambdas like I was expecting. `sam build` uses ESBuild under the hood to
transpile Typescript into Javascript and it's possible that was already
ignoring the test files. In either case this change does make our config
more similar to other teams on the programme.