GUA-432: Trigger alarms when the DLQs receive a message
Set up a Cloudwatch alarm for each of the dead letter queues that's
triggered when there's at least one message unprocessed in the queue.

The `ApproximateNumberOfMessagesVisible` metric seemed the most
appropriate of those available [[1]. It's an approximation, but it'll
always be greater than 0 if there's at least one message which is fine
for this usecase where we don't care how many messages there are.

There was also `NumberOfMessagesSent` which is "The number of messages
added to a queue" but I wasn't sure if that would also be triggered if
the whole Lambda failed to run and AWS put the message onto the DLQ,
rather than us calling `sendMessage` from inside the lambdas.

[1]: https://docs.aws.amazon.com/AWSSimpleQueueService/latest/SQSDeveloperGuide/sqs-available-cloudwatch-metrics.html