GUA-432: Increase message retention time on DLQs
Up to the maximum allowed 14 days, from the default of 4 days.

If a message does get put one of these queues, something has gone wrong.

4 days isn't a lot of time to work out a fix, especially over a bank
holiday weekend or a busy period like Christmas. Increasing the message
retention means that we have more time to make a fix and reprocess the
stored messages.