BAU: Refactor handler to match other lambdas
Since writing this first lambda function we've worked out how to use
`Promise.all()` and `Array.map()` to iterate over all the records in an
event body. This functional programming approach is more in the spirit
of the Airbnb style guide and means we can drop the 'no await in loop'
rule exemption.

Refactor the handler function here to use the same method as the others.