BAU: Update Architectural Diagram to reflect new plan
After starting work on [1] and [2] we realised the passthrough lambdas
weren't really offering us much value, so we've removed them here.

The theory initially was that we'd want an SQS queue, and a scalable
lambda allow us to efficently write to the SNS topic. We don't think
that's true anymore. We should be able to write directly from the
fargate containers, which themselves can scale horizontally, and we
don't expect to hit a bottleneck here.

So let's remove a few components and simplify things a bit

[1]: https://github.com/alphagov/di-account-management-backend/pull/36
[2]: https://github.com/alphagov/di-account-management-backend/pull/38