GUA-502: Change function names to match log groups
There's no direct way to get a Lambda function to use a pre-created log
group. Instead it'll always log to the group called
`/aws/lambda/${LambdaFunctionName}`. If that group doesn't already
exist, it'll create it.

This commit changes the function names to match what we want to appear
in the log group. It also makes sure the log groups are created before
the Lambda functions. This means the log groups are created and managed
by CloudFormation and use our retention settings, encryption key etc.

If we don't force the log groups to be created first, they'll get
created when the Lambdas are made, and use default settings.