GUA-512: Send lambda logs to Splunk
Sets up log subscription filters that send application logs from the
lambdas to Splunk.

These aren't created in the `dev` AWS account. We don't have a CSLS
config for `dev` and it doesn't make sense to create one since we share
the non-production index with all environments and teams on DI. We don't
want to accidentally fill that index with logs from our development
environment while we're testing things out.