GUA-467: Add lambda invoke permission
After testing this deployment it was clear that the function was not
invoking correctly.

With our SQS queues we can add the SQS as a direct trigger under events.
You can do this too with an EventSource[1] object of Type SNS with SNS
properties [2]

AWS docs then say:

> SAM generates AWS::SNS::Subscription resource when this event type is set

However we want a RedrivePolicy applied, which doesn't seem possible to
supply as an EventSource and have translated into an
AWS::SNS:Subscription, unless i missed something.

The solution is to use the existing AWS::SNS::Subscription, but that
seems to not be knitting together nicely for the lack of an AWS lambda
invoke permission [3].

So let's give that a go.

[1]: https://docs.aws.amazon.com/serverless-application-model/latest/developerguide/sam-property-function-eventsource.html
[2]: https://docs.aws.amazon.com/serverless-application-model/latest/developerguide/sam-property-function-sns.html
[3]: https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-lambda-permission.html