FIX: Tests for the raw-event processing in invalid data structure
In the function we're looking to test the event body directly, but here
we were encoding these into an object with a key of `services` and then
expecting these to be an array body.

This must have been missed in pass reviews.

In this case it means that we're passing tests becuase they _are_
throwing an error, but not becuase it's the right error, instead because
it can't find the `user.user_id` object.

Change with this adjustment, which reveals another issue with the tests
I'll fix in the next commit