FIX: tests incorrectly used clientId instead of client_id
Bit confusing as good JS syntax is to use camelCase instead of
snake_case, but this is how the data structure will be arriving with us
from TxMA and AWS land.

In this case the error was masked by another error where we were
throwing but now from an invaid TxMA Event Body, instead from an invalid
user payload. Alas, but easily fixed.