GUA-748: Remove models.ts from query-activity-log package.json
We're going to try something different this time around. In all our
other lambdas we have a models.ts file for each lambda in each lambda
folder. This is now resulting in an insanely boring duplication of code,
when we change one value we have to update in a lot of places.

We did this on the understanding of the serverless best pratices that
things should be kept local and you shouldn't have big shared folders.

However after reviewing we think that this shouldn't really apply to
models. Mostly becuase they disappear after typescript compiles, and
there's risk here around changing our underlying expectations of data
shapes and not seeeing that reflected in compiler errors becuase we have
different interface defintions elsewhere.

So we're gonna try a shared folder and remove this. that will be done on
another PR and pulled into this