GUA-748: Add a Query lambda for the Activity Log
This lambda reads from the dynamoDb stream and queries the Activity Log Dynamo DB with the user_id as the partition key .
The query filters based on the session_id .
Lambda has Validation to validate the TXMA events.
It sends a message on the DLQ in case of invalid data.