GUA-872: Add requirement to write mock event to SQS
Our OIDC mock needs to be two things - the OIDC server and the
event source that the backend consumes from. We should make the
stub as close to reality as possible, so the `/authorize` endpoint
is the one which should generate that event.

This does overload the purpose of an OpenAPI specification slightly;
they're meant to document the interface, not the workings of the API.
I didn't want to add another file though, and I don't think this is
too bad as it's only a small stub.