GUA-873: Add ADR for stub architecture
We will be running performance (load) testing to see how our service scales.
As part of this, we need to replace external dependencies with stubs.
These stubs will be simple interfaces to replace eg. the Account Management API.

This ADR sets out options for the hosting and architecture of the stubs.