GUA-874: Create lambda for AM API stub
This lambda responds to API Gateway events and always returns
204 status code and an empty body.

I had hoped to build this API stub completely with API Gateway
mock integrations, but they're only supported on API Gateway v1
for HTTP APIs. V1 doesn't have default routes or a fixed 'live' URL
so we'd need ~100 lines of Cloudformation per stubbed API route to
define the route, method, integration and response template.

Instead I thought it'd be easier to understand if we had a single
lambda which generates the response then use API Gateway v2 and a
default route to forward all requests to the lambda. This'll mean
we end up with fewer lines of code and it's still a permitted
architecture in [the ADR](https://github.com/alphagov/di-account-management-backend/blob/main/docs/adr/0007-architecture-for-stubs.md).