BAU: Reduce raw events TTL to 14 days
This is one of the mitigations from our threat modelling session
to reduce risk of data exposure in the raw event store. The system
has been happily running for several months now and we're not
going to need 90 days data to back fill anything.

14 days is still enough time for us to do any event replays if we
have an outage over a period like Christmas where we may have several
team members on leave.