GUA-874: Deploy Account Management API stub to build
This adds a new condition to use stubbed services in case we want
to use them in dev in the future too.

This setup is broadly the same as the example stack I attached to
PR #118 where I added the lambda. The only differences are the
logging and other production-ready stuff like the permissions
boundary and the VPCs.

I've intentionally not put a dead letter queue on this lambda
because it's a standalone thing returning a static payload. We'll
have enough info in the logs to debug any problems and it doesn't
matter if we 'lose' the event data that triggered an error.