BAU: Save successful ID check event
We're now receiving this event from TxMA which is triggered once a user
successfully completes the ID proving journey and is passed back to
Auth.

We'll want to show information to users about when they proved their ID
in the activity log, so add it to the list of allowed event names.

See https://github.com/alphagov/di-txma-event-self-serve/pull/231