Fix module mock
Move the module mock outside the Jest `describe`. This is necessary so
Jest can hoist the mock setup to before the actual import and create the
mock.

Then set up the mock as an ES6 module so we can use it for named
imports. See https://remarkablemark.org/blog/2018/06/28/jest-mock-default-named-export/

Now the tests fail on the validation steps like we were expecting