BAU: Consolidate Sonar tokens
It seems that at some point since we set up the Sonar scan actions Sonar
has changed their access token model. When we set it up we needed to use
one token per repository / project. These tokens were migrated from
project level to user access tokens based on whoever set up the project
initially.

This meant that when the person who set up a lot of these actions left
GDS all the tokens were invalidated, preventing us from merging any PRs.

As a stopgap, I've replaced all the tokens with a user token from my
account. I've already switched all the secret values to that new token
and verified it works for all the projects. This means we can point all
of the actions to the same secret.

Obviously this isn't ideal as now my account is the single point of
failure, but at least we only have one place to update in the future.