INCIDEN-657: Don't log full SQS message
This lambda was deployed to production for ~45 minutes before we pulled
it. During that time it processed 281 messages, errored on each one and
logged the full message to Cloudwatch when passing it on to the DLQ.
These messages contained user IDs which shouldn't be sent to Cloudwatch
or Splunk.

Remove the line which logs the full message.

This was likely left in while doing testing and can be safely
removed. We already have error handling and sensible log messages for
when something goes wrong which include the message ID if we need to
retrieve the message later.