OLH-1410 Change the primary key for the activity log table
Rename the activity_log table (from activity_logs to activity_log), this is because DynamoDB key schemas can't be changed after they have been created. By renaming the table, it deletes the old one and creates a new one (there is no data in the table so no need for moving the data over)

Rename the name for the table name parameter in the CloudFormation template. This is because it is a bit tricky to change a parameter value in CloudFormation once it has been created, so by using a new one, we can avoid having to change the value manually.

Update the lambdas that used the old timestamp property to use event_id

A few minor improvements to the codebase that I spotted.