OLH-1485: Increase message visibility for audit events
TxMA are changing the timeout on their event processing pipeline to 60
seconds. Before they can do that, all event producers need to increase
the message visibility on their queues to reflect the new timeout (see TT2-1450).

Change the timeout on our queue from the default of 30 seconds to the
requested 70 seconds.