OLH-1562: Add Home service name for emails
We'll be showing sign in events for One Login Home in the activity
history. Information about Home isn't in the mapping here because it was
copied over from the service card data in the frontend and we don't have
a card for Home.

Add just the service info in English and Welsh so we can reference it in
the emails we send.

This uses the real client IDs for staging, integration and production so
we have them available for end to end testing