OLH-1562: Add Home service name for emails (#208)
## Proposed changes
### What changed

Add the service info for One Login Home in English and Welsh so we can
reference it in the emails we send.

This uses the real client IDs for staging, integration and production so
we have them available for end to end testing.

### Why did it change

We'll be showing sign in events for One Login Home in the activity
history. Information about Home isn't in the mapping here because it was
copied over from the service card data in the frontend and we don't have
a card for Home.

## Checklists

### Environment variables or secrets

- [x] No environment variables or secrets were added or changed

## Testing

I plan to do a full test of this once I've done the corresponding
frontend work to display the Home events.