OLH-1605 - Move user email into custom field/body and remove from Requestor (#215)
## Proposed changes

OLH-1605 - Move user email into custom field/body and remove from
Requestor

### What changed

Code to add the requester email as part of the message body and remove
the requester field.

### Why did it change

So that TICF still has a way to filter tickets by user, however the
generic email sent to new Zendesk users is not triggered. This is needed
because the Zendesk configurations required to stop the email generic
email from sending did not work without some caveats.