BAU: Trigger the Cloudformation lint action correctly
Fix the trigger conditions on the CFN lint actions so:
1. We run the linter on a pull request whenever `deploy.yaml` is changed
2. We run the 'skip action' on all other pull requests

This must have been broken since we refactored this repo into a single
package but we didn't notice because we had the skip action always
running on every pull request, which always passes.

This meant that we merged #260 with a malformed Cloudformation template
and broke our pipeline. That was addressed separately in #261.