OLH-1934: Give Orchestration permission to publish
Allow lambdas in Orchestration's AWS accounts to publish to our deletion
SNS topic.

They're working on a new process for manually deleting accounts if a
user requests us to do so through eg. the contact centre. In that
process the 2nd line developer will trigger a lambda in their account
which will delete the data from their stores and then post the relevant
identifiers to our SNS topic, allowing the deletion to cascade across
the many services in the programme.