OLH-1949: Add a script to invoke lambdas with a fixture
This script uses the AWS cli to invoke the `write-activity-log` and
`delete-activity-log` lambdas with pre-defined fixtures. It'll fail with
a non-zero exit code if either invocation fails.

This is a super-simple integration test, but it checks a lot:
- The lambdas have built correctly and can run
- They have permission to use KMS and DynamoDB
- They can process the message without timing out

Notably this script won't fail if any of the payloads are invalid due to
the way we catch and handle errors in the lambdas. That's ok for now.

In the future we'll probably want to migrate this to an NPM script using
something like Cucumber to orchestrate the tests. This is a first go at
getting a post-deploy test working; it'll catch the most broken bad
releases as a start.