OLH-1949: Allow the test role to invoke lambdas
The default role the post-deploy tests run as doesn't have many
permissions. The pipeline provides the `TestRoleArn` parameter to the
deployed stack so we can build our own policy and grant it the
permissions we need.

Add this policy to our template and allow the role to invoke the two
lambdas that we want to test