OLH-1934: Give Orchestration permission to publish (#263)
## Proposed changes

### What changed

Allow lambdas in Orchestration's AWS accounts to publish to our deletion
SNS topic. This also means they need permissions for the KMS key that
encrypts that topic.

I've copied these permissions from the frontend which also publishes to
this topic so I'm reasonably confident they're correct.

### Why did it change

They're working on a new process for manually deleting accounts if a
user requests us to do so through eg. the contact centre. In that
process the 2nd line developer will trigger a lambda in their account
which will delete the data from their stores and then post the relevant
identifiers to our SNS topic, allowing the deletion to cascade across
the many services in the programme.

### Related links


## Checklists

<!-- Merging this PR deploys to production. Please answer accurately.
-->

### Environment variables or secrets

- [x] No environment variables or secrets were added or changed

### Permissions

- [x] This PR adds or changes permissions


## How to review

Please check the account IDs here against the ones on the Jira ticket.