OLH-1934: Don't give Orchestration permissions in build
We only have account IDs from Orchestration in staging, integration and
production environments. This is correct, as those are the only
environments that get events from TxMA and therefore where data needs to
be deleted.

However, the condition on this policy was to add it in build as well.
This caused the deploy to fail as the principal ends up malformed.

Add a new condition to be clear when we're using external audit events,
and use that to only enable the SNS and KMS policies in the correct
environments.