OLH-1949: Provide the role name to the test policy
The pipeline provides us with the test container role ARN, but to attach
a managed policy to an existing role we need the role name instead.

Thankfully ARNs are a predictable format and are always like

`arn:aws:iam::123456789:role/role-name`

Everything after the `/` is the role name. Use the Cloudformation built
in functions to split the ARN and select the second item in the array
which is the name.