OLH-1949: Copy test fixtures to Dockerfile root
I can run the test container in the pipeline but it fails with
`Unable to load paramfile file://post-deploy-tests/fixtures/write-activity-log.json: [Errno 2] No such file or directory:`

I think this is to do with the working directory and where the script is
run from. I also realised we don't need to copy the whole repo into the
container - we only need the test script and the fixtures for this.

Switch to only copying the files we need and use relative file paths for
the fixtures so hopefully they can be found when run in the pipeline.