OLH-1949: Use `cat` to read test fixtures
We are still seeing `[Errno 2] No such file or directory` when the test
container runs in the pipeline but not locally.

The `file://` URIs are notoriously picky so stop using them completely
and switch to using `cat` to read the files in place. This means we need
to run the script with bash and not plain sh. I've checked and the
Amazon Linux image does include bash.