OLH-1949: Refactor post-deploy test script
We're using the same AWS CLI command to perform the same invocation and
same validation step on both lambdas.

Extract the testing out into a function and then call the new function
iterating over all the lambda names.

This will make it easier and cleaner to add new lambdas in the future if
we want to - we'll just need to add a fixture and then update the array
on line 18.