OLH-1949: Use `cat` to read test fixtures (#305)
## Proposed changes

<!-- Provide a summary of your changes in the title above -->
<!-- Include the Jira ticket number in square brackets as prefix, eg
`[OLH-XXXX]: PR Title` -->

### What changed

Use `cat` to read test fixtures in the post-deploy test script. 

### Why did it change

We are still seeing `[Errno 2] No such file or directory` when the test
container runs in the pipeline but not locally.

The `file://` URIs are notoriously picky so stop using them completely
and switch to using `cat` to read the files in place. This means we need
to run the script with bash and not plain sh. I've checked and the
Amazon Linux image does include bash.

## Checklists

<!-- Merging this PR deploys to production. Please answer accurately.
-->

### Environment variables or secrets

- [x] No environment variables or secrets were added or changed

## Testing

I can build and run the container locally and the pipeline step is still
disabled. Once this is merged I'll go back and re-try it.