OLH-1949: Validate lambda invocations in the post-deploy tests (#306)
## Proposed changes

### What changed

The response we get back from the Lambda service looks like:
```json
{
    "StatusCode": 200,
    "ExecutedVersion": "60"
}
```

Use JQ to parse this response and check the status code was 200. If it's
not, halt processing with exit code 1. This will propagate up and cause
the script to exit, failing the tests in Codebuild.

Also increase the `write-activity-log` fixture to 10 messages. This was
the original plan for these tests as we discovered that the
`write-activity-log` lambda timed out when it processed a full batch of
10 messages on the default timeout settings.

Don't merge this until #305 has been merged into `main`. 

### Why did it change

At the moment any errors raised by these tests will be to do with the
test setup - either in the script or AWS permissions related. We also
need to check the lambda responses and make sure the function succeeded
for this to be a useful test suite.

## Checklists

### Environment variables or secrets

- [x] No environment variables or secrets were added or changed

## Testing

I've built the test container locally and run it against build,
following the instructions in the README.