OLH-2293-delete-email-subscription-lambda-failing-in-production (#442)
## Proposed changes

<!-- Provide a summary of your changes in the title above -->
<!-- Include the Jira ticket number in square brackets as prefix, eg
`[OLH-XXXX]: PR Title` -->
[OLH-2293] Resolve Delete Email Subscription not picking up latest
secret value from Secrets Manager

### What changed

<!-- Describe the changes in detail - the "what"-->
Implemented Version ID for Publishing API Token. If the key changes in
the future the version id will need to be updated in cloud formation,
triggering the code deploy that will actually get the new secret from
secret's manager.

### Why did it change
<!-- Describe the reason these changes were made - the "why" -->
Cloud formation did not update the secret value in the lambda on
deployment as the section of code that check's did not change. This is
as cloud formation only updates stuff that changes in the template.yaml
file. Moving forward we will be forced to deploy an updated version ID
which will trigger the Lambda to have an updated value. NOTE: We should
not manually update the value in the environments as this will break
cloud formation. We opted not to get the secret's value within the
Lambda as it will lead to excessive calls for a value that changes ~2-3
years.

## Checklists

<!-- Merging this PR deploys to production. Please answer accurately.
-->

### Environment variables or secrets

- [ ] No environment variables or secrets were added or changed

<!-- Delete if changes DO NOT include new environment variables or
secrets -->

- [x] Added parameters to Cloudformation template
- [x] Added new environment variable

## Testing

<!-- Provide a summary of any manual testing you've done, for example
deploying the branch to dev -->
Deployed to Dev


[OLH-2293]:
https://govukverify.atlassian.net/browse/OLH-2293?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ